pyproject.toml
[project]
name = "code-sage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "argparse>=1.4.0",
    "chromadb>=1.2.2",
    "fastapi>=0.120.1",
    "gitpython>=3.1.45",
    "rich>=14.2.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
]
